handson/05-bashoutter/requirements.txt
aws-cdk.aws-s3==1.30.0
aws-cdk.aws-s3-deployment==1.30.0
aws-cdk.aws-dynamodb==1.30.0
aws-cdk.aws-lambda==1.30.0
aws-cdk.aws-apigateway==1.30.0
boto3
requests
httpie
